remove property name from list
Property name was in the original list to take a URI of a previously defined term, so you could say that, for example, you would use "dct:title" as the URI for your statement. We will almost surely need to add this back in, but it shouldn't be needed for the hack day experiments.